+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.4] - 2017-08-30
 
 ### Added
 
@@ -94,7 +94,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Simple library for using the functionality in other projects
 - Web interface for downloading youtube videos
 
-[Unreleased]: https://github.com/jeckman/YouTube-Downloader/compare/0.3...HEAD
+[Unreleased]: https://github.com/jeckman/YouTube-Downloader/compare/0.4...HEAD
+[0.4]: https://github.com/jeckman/YouTube-Downloader/compare/0.3...0.4
 [0.3]: https://github.com/jeckman/YouTube-Downloader/compare/0.2...0.3
 [0.2]: https://github.com/jeckman/YouTube-Downloader/compare/0.1...0.2
 [0.1]: https://github.com/jeckman/YouTube-Downloader/compare/7397d4101a96aa3cc28d211d9f23d5da34c3d9c8...0.1